ISE-308
renamed the mkdirs cookbook to satisfy the naming standards of our
private cookbooks. Removed the mysql-nativex cookbook as it was not a
production cookbook and was early phase test that should have been
deleted long ago. Removed all references to the mysql community
cookbooks as well as the yum-mysql comunity cookbook.